+++ app/Http/Requests/StoreDeadlineRequest.php
@@ -2,17 +2,12 @@
 
 namespace App\Http\Requests;
 
+use App\Http\Requests\Concerns\AdminOnlyAccess;
 use Illuminate\Foundation\Http\FormRequest;
 
 class StoreDeadlineRequest extends FormRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
+    use AdminOnlyAccess;
 
     /**
      * Get the validation rules that apply to the request.