+++ app/Http/Requests/StoreVehicleRequest.php
@@ -3,18 +3,12 @@
 namespace App\Http\Requests;
 
 use App\Http\Requests\Concerns\HandlesWarrantyExtension;
+use App\Http\Requests\Concerns\AdminOnlyAccess;
 use Illuminate\Foundation\Http\FormRequest;
 
 class StoreVehicleRequest extends FormRequest
 {
-    use HandlesWarrantyExtension;
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
+    use HandlesWarrantyExtension, AdminOnlyAccess;
 
     /**
      * Get the validation rules that apply to the request.