+++ app/Http/Requests/UpdateEquipmentRequest.php
@@ -2,19 +2,14 @@
 
 namespace App\Http\Requests;
 
+use App\Http\Requests\Concerns\AdminOnlyAccess;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Carbon;
 use Illuminate\Validation\Validator;
 
 class UpdateEquipmentRequest extends FormRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
+    use AdminOnlyAccess;
 
     /**
      * Get the validation rules that apply to the request.