+++ app/Http/Requests/UpdateVehicleRequest.php
@@ -2,20 +2,14 @@
 
 namespace App\Http\Requests;
 
+use App\Http\Requests\Concerns\AdminOnlyAccess;
 use App\Http\Requests\Concerns\HandlesWarrantyExtension;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 
 class UpdateVehicleRequest extends FormRequest
 {
-    use HandlesWarrantyExtension;
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
+    use HandlesWarrantyExtension, AdminOnlyAccess;
 
     /**
      * Get the validation rules that apply to the request.