+++ app/Policies/ProviderPolicy.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Policies;
+
+use App\Models\Provider;
+use App\Policies\Concerns\HasRoleBasedAccess;
+
+class ProviderPolicy
+{
+    use HasRoleBasedAccess;
+}